Adding FusionCharts borders & background

https://www.fusioncharts.com/dev/chart-guide/chart-configurations/border-and-background

Usage

fusionBackground(
  fusionPlot,
  showBorder = FALSE,
  borderColor = "#666666",
  borderThickness = "4",
  borderAlpha = "80",
  bgColorStart = "#ffffff",
  bgColorEnd = NULL,
  bgAlphaStart = "50",
  bgAlphaEnd = NULL,
  bgratioStart = "60",
  bgratioEnd = "40",
  bgAngle = "180"
)

Arguments

fusionPlot

fusionPlot object got by fusionPlot()

showBorder

Show the chart border

borderColor

Specify the color of the border

borderThickness

Set the thickness of the border

borderAlpha

Set the transparency of the border

bgColorStart, bgColorEnd

Set the hex codes of the starting and ending gradient colors

bgAlphaStart, bgAlphaEnd

Set the transparency of the starting ending gradient colors

bgratioStart, bgratioEnd

Set the radius of gradient colors

bgAngle

Set the angle in degrees of gradient colors

Examples

library(fusionchartsR)
df <- data.frame(label = c("Venezuela", "Saudi", "Canada", "Russia"), value = c(290, 260,180, 115))
df %>%
fusionPlot(x = "label", y = "value", type = "column3d") %>%
fusionBackground(showBorder = TRUE, bgColorStart = "#DDDDDD") %>%
fusionTheme(theme = "fusion")