Customing FusionCharts axis

https://www.fusioncharts.com/dev/chart-guide/chart-configurations/axes

Usage

fusionCustomAxis(
  fusionPlot,
  showlabels = TRUE,
  xAxisPosition = c("bottom", "top", "left", "right"),
  yAxisPosition = c("left", "right", "top", "bottom"),
  yAxisMinValue = NULL,
  yAxisMaxValue = NULL,
  AxisNameBorderColor = NULL,
  AxisNameBorderAlpha = "0",
  AxisNameBorderPadding = "6",
  AxisNameBorderRadius = "3",
  AxisNameBorderThickness = "2",
  AxisNameBorderDashed = FALSE,
  AxisNameBorderDashLen = "4",
  AxisNameBorderDashGap = "2",
  AxisNameBgColor = NULL,
  AxisNameBgAlpha = "0",
  AxisNameFontAlpha = "100",
  AxisValueFont = "Arial",
  AxisValueFontSize = "1px",
  AxisValueFontColor = NULL,
  AxisValueFontBold = FALSE,
  AxisValueFontItalic = FALSE,
  AxisValueAlpha = "100",
  AxisValueBgColor = NULL,
  AxisValueBgAlpha = "50",
  AxisValueBorderColor = "#ffffff",
  AxisValueBorderAlpha = "0",
  AxisValueBorderPadding = "5",
  AxisValueBorderRadius = "2",
  AxisValueBorderThickness = "3",
  AxisValueBorderDashed = FALSE,
  AxisValueBorderDashLen = "2",
  AxisValueBorderDashGap = "2"
)

Arguments

fusionPlot

fusionPlot object got by fusionPlot()

showlabels

Display the data labels

xAxisPosition

change the position of the x-axis

yAxisPosition

change the position of the y-axis

yAxisMinValue

Set the lower limit of the x-axis

yAxisMaxValue

Set the upper limit of the y-axis

AxisNameBorderColor

Set the border color of the name of the axis

AxisNameBorderAlpha

Set the transparency of the border around the name of axis

AxisNameBorderPadding

Set the padding of the border around the name of the axis

AxisNameBorderRadius

Set the radius of the border around the name of the axis

AxisNameBorderThickness

Set the thickness of the border around the name of the axis

AxisNameBorderDashed

Make the border around the name of the axis dashed

AxisNameBorderDashLen

Set the length of each dash in the dashed border around the name of the axis

AxisNameBorderDashGap

Set the gap between two consecutive dashes in the dashed border around the name of the axis

AxisNameBgColor

Set the background color of the name of the axis

AxisNameBgAlpha

Set the transparency of the background of the name of the axis

AxisNameFontAlpha

Set the transparency of the name of the axis

AxisValueFont

Set the font of the axis values

AxisValueFontSize

Set the font size (between 0 to 72) of the axis values

AxisValueFontColor

Set the font color of the axis

AxisValueFontBold

Set the font of the axis values to bold

AxisValueFontItalic

Set the font for the axis values to italics

AxisValueAlpha

Set the degree of transparency of the axis values

AxisValueBgColor

Set the background color of the axis values

AxisValueBgAlpha

Set the background color transparency of the axis values

AxisValueBorderColor

Set the border color of the axis values

AxisValueBorderAlpha

Set the transparency of the border of the axis values

AxisValueBorderPadding

Set the padding of the axis values border

AxisValueBorderRadius

Set the border radius of the axis values

AxisValueBorderThickness

Set the border thickness of the axis values

AxisValueBorderDashed

Make the axis values border dashed

AxisValueBorderDashLen

Set the length of each dash for the dashed borders around axis values

AxisValueBorderDashGap

Set the gap between two consecutive dashes for the dashed borders around the axis values

Examples

library(fusionchartsR)
df <- data.frame(label = c("Venezuela", "Saudi", "Canada", "Russia"), value = c(290, 260, 180, 115))
df %>%
  fusionPlot(x = "label", y = "value", type = "column3d") %>%
  fusionAxis(xAxisName = "Countries", yAxisName = "Numbers", AxisNameFontSize = "20") %>%
  fusionCustomAxis(xAxisPosition = "top", yAxisPosition = "right") %>%
  fusionTheme(theme = "gammel")