Adding FusionCharts palette

https://www.fusioncharts.com/dev/chart-guide/chart-configurations/data-plot

Usage

fusionPalette(
  fusionPlot,
  palettecolors = NULL,
  usePlotGradientColor = FALSE,
  plotGradientColor = "#003366",
  plotFillAngle = "0",
  plotFillRatioStart = "90",
  plotFillRatioEnd = "100",
  plotFillAlpha = "100",
  showPlotBorder = FALSE,
  drawFullAreaBorder = FALSE,
  inheritPlotBorderColor = FALSE,
  plotBorderDashed = FALSE,
  plotBorderDashLen = "4",
  plotBorderDashGap = "4",
  plotBorderThickness = "1",
  plotBorderColor = "#666666",
  useRoundEdges = FALSE,
  plotHoverEffect = FALSE,
  plotFillHoverColor = "#5D62B5",
  plotFillHoverAlpha = "100",
  plotBorderHoverColor = "#000000",
  plotBorderHoverAlpha = "100",
  plotBorderHoverThickness = "1",
  plotBorderHoverDashed = TRUE,
  plotBorderHoverDashLen = "6",
  plotBorderHoverDashGap = "2"
)

Arguments

fusionPlot

fusionPlot object got by fusionPlot()

palettecolors

Specify your custom palette for data plots

usePlotGradientColor

Use the gradient effect

plotGradientColor

Specify the hex code of the gradient color

plotFillAngle

Set the fill angle for the gradient (0 to 360)

plotFillRatioStart

Specify the start of the gradient effect

plotFillRatioEnd

Specify the end of the gradient effect

plotFillAlpha

Set the transparency of the gradient fill

showPlotBorder

Show the plot border

drawFullAreaBorder

To set the top border of the area chart (only works if showPlotBorder = TRUE)

inheritPlotBorderColor

Enable the plot border to inherit the color of an area plot

plotBorderDashed

Make the border dashed

plotBorderDashLen

Set the length of each dash in plot-border (in pixels)

plotBorderDashGap

Set the gap between two consecutive dashes in plot border (in pixels)

plotBorderThickness

Set the thickness of the plot border

plotBorderColor

Set the color of the plot border

useRoundEdges

Enable rounded edges (2D Column or Bar charts only)

plotHoverEffect

Enable hover effects for the data plots

plotFillHoverColor

Set the hover color for data plots in hex code format

plotFillHoverAlpha

Set the transparency for hover color for data plots

plotBorderHoverColor

Set the hover border color

plotBorderHoverAlpha

Set the transparency of hover border for data plots

plotBorderHoverThickness

Set the hover border thickness (in pixels)

plotBorderHoverDashed

Make dashed borders on hover

plotBorderHoverDashLen

Set the length of each dash for all data plots on hover

plotBorderHoverDashGap

Set the gap between two consecutive dashes for all data plots on hover(in pixels)

Examples

library(fusionchartsR)
df <- data.frame(label = c("Venezuela", "Saudi", "Canada", "Russia"), value = c(290, 260,180, 115))

df %>%
fusionPlot(x = "label", y = "value", type = "pie2d") %>%
fusionPalette(palettecolors = c("5d62b5", "29c3be", "f2726f")) %>%
fusionTheme(theme = "gammel")